Feat: test/JBETHPaymentTerminal (#49)
* Feat: init

* Feat: so far, so good - returns balance...

* cuz I fell in love with the girl at the mock show

* wip

* feat: wip

* wip

* wip

* wip

* wip

* wip

* - stuck -

* - stuck -

* - stuck/dirty trick -

* wip

* wip

* useAllowanceOf tests

* wip dpo

* wip/getting there

* fix: update tracker

* wip

* feat: distribute payout of if fucking done (I guess)

* feat: migrate

* feat: add to balance of (to refactor+++)

* Correct the bit packing for recordPaymentFrom call

* redeemTokensOf tests

* delegate and processFees tests

* feat: setFee test

* fix: .only

* feat: coverage 100-100-100

* chore: cleaning

* comments for context

* fix: test crossterminal/same terminal in distribute

* fix: test for event Pay

* fix: review

* fix: max split percent hardcoded

* feat: helpers reorg

+++ contracts/JBETHPaymentTerminal.sol
@@ -99,9 +99,9 @@ contract JBETHPaymentTerminal is
   */
   JBETHPaymentTerminalStore public immutable store;
 
-  /** 
-    @notice 
-    The token that this terminal accepts. 
+  /**
+    @notice
+    The token that this terminal accepts.
   */
   address public immutable override token = JBTokens.ETH;
 
@@ -364,9 +364,8 @@ contract JBETHPaymentTerminal is
     bytes32 _handle = projects.handleOf(_projectId);
 
     // Take a fee from the _withdrawnAmount, if needed.
-    // The project's owner will be the beneficiary of the resulting minted tokens from platform project.
-    // The platform project's ID is 1.
-    uint256 _feeAmount = fee == 0 || _projectId == 1
+    // The project's owner will be the beneficiary.
+    uint256 _feeAmount = fee == 0 || _projectId == 1 // The platform project's ID is 1.
       ? 0
       : _takeFeeFrom(
         _projectId,
@@ -700,8 +699,8 @@ contract JBETHPaymentTerminal is
     if (feeAmount == 0) return 0;
 
     _fundingCycle.shouldHoldFees()
-      ? _heldFeesOf[_projectId].push(JBFee(_amount, uint8(fee), _beneficiary, _memo)) // Take the fee.
-      : _takeFee(feeAmount, _beneficiary, _memo);
+      ? _heldFeesOf[_projectId].push(JBFee(_amount, uint8(fee), _beneficiary, _memo))
+      : _takeFee(feeAmount, _beneficiary, _memo); // Take the fee.
   }
 
   /**
@@ -721,9 +720,9 @@ contract JBETHPaymentTerminal is
     IJBTerminal _terminal = directory.primaryTerminalOf(1, token);
 
     // When processing the admin fee, save gas if the admin is using this contract as its terminal.
-    _terminal == this // Use the local pay call.
-      ? _pay(_amount, 1, _beneficiary, 0, false, _memo, bytes('')) // Use the external pay call of the correct terminal.
-      : _terminal.pay{value: _amount}(1, _beneficiary, 0, false, _memo, bytes(''));
+    _terminal == this
+      ? _pay(_amount, 1, _beneficiary, 0, false, _memo, bytes('')) // Use the local pay call.
+      : _terminal.pay{value: _amount}(1, _beneficiary, 0, false, _memo, bytes('')); // Use the external pay call of the correct terminal.
   }
 
   /**
@@ -753,7 +752,7 @@ contract JBETHPaymentTerminal is
       msg.sender,
       _amount,
       _projectId,
-      (_preferClaimedTokens ? 1 : 0) | (uint160(_beneficiary) << 1),
+      (_preferClaimedTokens ? 1 : 0) | (uint256(uint160(_beneficiary)) << 1),
       _minReturnedTokens,
       _memo,
       _delegateMetadata
